+++ cf-monitord/Makefile.am
@@ -3,6 +3,7 @@ noinst_LTLIBRARIES = libcf-monitord.la
 AM_CPPFLAGS = @CPPFLAGS@ \
 	-I$(srcdir)/../libutils \
 	-I$(srcdir)/../libcfnet \
+	-I$(srcdir)/../libenv \
 	-I$(srcdir)/../libpromises \
 	$(ENTERPRISE_CPPFLAGS)
