+++ cf-monitord/mon_network_sniffer.c
@@ -24,7 +24,7 @@
 
 #include <cf3.defs.h>
 
-#include <sysinfo.h>
+//#include <sysinfo.h>
 #include <files_names.h>
 #include <files_interfaces.h>
 #include <mon.h>
@@ -205,7 +205,23 @@ static void IncrementCounter(Item **list, char *name)
     IncrementItemListCounter(*list, name);
 }
 
-/******************************************************************************/
+static bool IsInterfaceAddress(const char *adr)
+ /* Does this address belong to a local interface */
+{
+    Item *ip;
+
+    for (ip = IPADDRESSES; ip != NULL; ip = ip->next)
+    {
+        if (strncasecmp(adr, ip->name, strlen(adr)) == 0)
+        {
+            Log(LOG_LEVEL_DEBUG, "Identifying '%s' as one of my interfaces", adr);
+            return true;
+        }
+    }
+
+    Log(LOG_LEVEL_DEBUG, "'%s' is not one of my interfaces", adr);
+    return false;
+}
 
 /* This coarsely classifies TCP dump data */
