+++ cf-serverd/Makefile.am
@@ -2,6 +2,7 @@ noinst_LTLIBRARIES = libcf-serverd.la
 
 AM_CPPFLAGS = @CPPFLAGS@ -I$(srcdir)/../libpromises -I$(srcdir)/../libutils \
 	-I$(srcdir)/../libcfnet \
+	-I$(srcdir)/../libenv \
 	$(ENTERPRISE_CPPFLAGS)
 
 AM_CFLAGS = @CFLAGS@ \
