+++ cf-serverd/cf-serverd-functions.c
@@ -37,7 +37,10 @@
 #include <man.h>
 #include <tls_server.h>                              /* ServerTLSInitialize */
 #include <timeout.h>
-
+#include <unix_iface.h>
+#include <known_dirs.h>
+#include <sysinfo.h>
+#include <time_classes.h>
 
 static const size_t QUEUESIZE = 50;
 int NO_FORK = false;
