+++ libenv/sysinfo.c
@@ -39,6 +39,7 @@
 #include <rlist.h>
 #include <audit.h>
 #include <pipes.h>
+#include <known_dirs.h>
 
 #include <cf-windows-functions.h>
 
@@ -864,28 +865,6 @@ void Get3Environment(EvalContext *ctx, AgentType agent_type)
     fclose(fp);
     Log(LOG_LEVEL_VERBOSE, "Environment data loaded");
 }
-
-/*******************************************************************/
-
-_Bool IsInterfaceAddress(const char *adr)
- /* Does this address belong to a local interface */
-{
-    Item *ip;
-
-    for (ip = IPADDRESSES; ip != NULL; ip = ip->next)
-    {
-        if (strncasecmp(adr, ip->name, strlen(adr)) == 0)
-        {
-            Log(LOG_LEVEL_DEBUG, "Identifying '%s' as one of my interfaces", adr);
-            return true;
-        }
-    }
-
-    Log(LOG_LEVEL_DEBUG, "'%s' is not one of my interfaces", adr);
-    return false;
-}
-
-/*******************************************************************/
 
 void BuiltinClasses(EvalContext *ctx)
 {
@@ -2416,119 +2395,6 @@ static FILE *ReadFirstLine(const char *filename, char *buf, int bufsize)
     return fp;
 }
 #endif /* __linux__ */
-
-/******************************************************************/
-/* User info                                                      */
-/******************************************************************/
-
-#if defined(__CYGWIN__)
-
-static const char *GetDefaultWorkDir(void)
-{
-    return WORKDIR;
-}
-
-static const char *GetDefaultLogDir(void)
-{
-    return LOGDIR;
-}
-
-static const char *GetDefaultPidDir(void)
-{
-    return PIDDIR;
-}
-
-#elif defined(__ANDROID__)
-
-static const char *GetDefaultWorkDir(void)
-{
-    /* getpwuid() on Android returns /data, so use compile-time default instead */
-    return WORKDIR;
-}
-
-static const char *GetDefaultLogDir(void)
-{
-    return LOGDIR;
-}
-
-static const char *GetDefaultPidDir(void)
-{
-    return PIDDIR;
-}
-
-#elif !defined(__MINGW32__)
-
-#define MAX_WORKDIR_LENGTH (CF_BUFSIZE / 2)
-
-static const char *GetDefaultDir_helper(char dir[MAX_WORKDIR_LENGTH], const char *root_dir)
-{
-    if (getuid() > 0)
-    {
-        if (!*dir)
-        {
-            struct passwd *mpw = getpwuid(getuid());
-
-            if (snprintf(dir, MAX_WORKDIR_LENGTH, "%s/.cfagent", mpw->pw_dir) >= MAX_WORKDIR_LENGTH)
-            {
-                return NULL;
-            }
-        }
-        return dir;
-    }
-    else
-    {
-        return root_dir;
-    }
-}
-
-static const char *GetDefaultWorkDir(void)
-{
-    static char workdir[MAX_WORKDIR_LENGTH];
-    return GetDefaultDir_helper(workdir, WORKDIR);
-}
-
-static const char *GetDefaultLogDir(void)
-{
-    static char logdir[MAX_WORKDIR_LENGTH];
-    return GetDefaultDir_helper(logdir, LOGDIR);
-}
-
-static const char *GetDefaultPidDir(void)
-{
-    static char piddir[MAX_WORKDIR_LENGTH];
-    return GetDefaultDir_helper(piddir, PIDDIR);
-}
-
-#endif
-
-/******************************************************************/
-
-const char *GetWorkDir(void)
-{
-    const char *workdir = getenv("CFENGINE_TEST_OVERRIDE_WORKDIR");
-
-    return workdir == NULL ? GetDefaultWorkDir() : workdir;
-}
-
-/******************************************************************/
-
-const char *GetLogDir(void)
-{
-    const char *logdir = getenv("CFENGINE_TEST_OVERRIDE_LOGDIR");
-
-    return logdir == NULL ? GetDefaultLogDir() : logdir;
-}
-
-/******************************************************************/
-
-const char *GetPidDir(void)
-{
-    const char *piddir = getenv("CFENGINE_TEST_OVERRIDE_PIDDIR");
-
-    return piddir == NULL ? GetDefaultPidDir() : piddir;
-}
-
-/******************************************************************/
 
 static void GetCPUInfo(EvalContext *ctx)
 {
