+++ libpromises/generic_agent.c
@@ -26,6 +26,7 @@
 
 #include <bootstrap.h>
 #include <sysinfo.h>
+#include <known_dirs.h>
 #include <env_context.h>
 #include <policy.h>
 #include <promises.h>
@@ -56,6 +57,9 @@
 #include <verify_classes.h>
 #include <verify_vars.h>
 #include <timeout.h>
+#include <time_classes.h>
+#include <unix_iface.h>
+#include <constants.h>
 
 #include <cf-windows-functions.h>
