+++ libpromises/vars.c
@@ -37,17 +37,6 @@
 
 static int IsCf3Scalar(char *str);
 
-void LoadSystemConstants(EvalContext *ctx)
-{
-    EvalContextVariablePutSpecial(ctx, SPECIAL_SCOPE_CONST, "dollar", "$", DATA_TYPE_STRING);
-    EvalContextVariablePutSpecial(ctx, SPECIAL_SCOPE_CONST, "n", "\n", DATA_TYPE_STRING);
-    EvalContextVariablePutSpecial(ctx, SPECIAL_SCOPE_CONST, "r", "\r", DATA_TYPE_STRING);
-    EvalContextVariablePutSpecial(ctx, SPECIAL_SCOPE_CONST, "t", "\t", DATA_TYPE_STRING);
-    EvalContextVariablePutSpecial(ctx, SPECIAL_SCOPE_CONST, "endl", "\n", DATA_TYPE_STRING);
-/* NewScalar("const","0","\0",cf_str);  - this cannot work */
-
-}
-
 /*******************************************************************/
 
 int UnresolvedArgs(Rlist *args)
